Adding `Animation` object, and ability to chain animations together
The `Animation` object allows packaging a collection of `MotionGenerators`
that can all be run as a group. The animation is considered complete
once all motion generators are finished.

In addition to using the `Animation` object in `VehicleController` as a test
bed, adding the ability to chain multiple animations together. This allows
the "Y-rotation first, and then rest of movement" logic to be set via
configuration of the animations collection, rather than hard-coded logic.
This should make different types of animation in `VehicleController` easier
to implement. (E.g. by having an animation that doesn't require the Y-rotation
to happen separately from the other animations).