Improvements to intro animation for a flythrough
The animation is now guaranteed to take between 1 and 3 seconds. This
avoids agonizing slow intros when the camera is far away from the city
center, as well as intros that are too quick and sudden.

The Y-rotation toward the target is slower, and winds down to a gradual
stop. It also occurs at the same time as the rest of the animations. This
makes the animation less sudden and jarring.

The X-rotation toward the target also more quickly moves close to the
target position, and gradually slows to a stop.