Simplifying flythrough intro animation
Removing the "fly around city center in a circle" step. The intro should
be short and sweet, and get you into the flythrough with little fuss. More
complicated flythrough animations can occur as parts of the flythrough, but
not be required to happen at the beginning of each and every one.