Using more intuitive angles for quadrant detection
When determining the quadrant the camera is in, the Z position
is treated like the Y value in the call to `atan2()`. Since the
numbers on the Z-axis get larger as they get _closer_ to the viewer,
an uncorrected Z-axis will cause the angles retuned by `atan2()` to be
as if the Y-axis was upside down (i.e. negative numbers above axis,
positive numbers below).

The end results are the same as before, since the corresponding
quadrant boundaries have been updated as well. This seems like a more
intuitive way to think about the quadrant boundaries.