Adding inertial movement for center X/Z of interactive camera
That is, it will continue moving in the drag direction after the
mouse button/touch is released, until it slows to a stop.

Currently this is only enabled for the X/Z center position, but
the general pattern can be used for the Y-rotation, tilt, and zoom
parameters as well.