Simplifying `PoleCamera` implementation
The notion of a "pole" is now removed, so this now consists of just a camera,
rather than a "camera mounted on a pole".

Changing the rotation order so that Y-rotations happen before X-rotations
removes the need for the "camera mounted on a pole, so that camera pitch
doesn't affect the pole's heading" model. While a clever hack given my level
of knowledge at the time, it's simpler to just do everything using only a camera.

The name "PoleCamera" is not really accurate anymore, and this object arguably
no longer has a good reason to exist since it just wraps the camera object.
However, will deal with those issues in different commit(s).