InteractiveCamera internally tracks zoom distance, tilt angle
Previously, it only tracked the percentage of the allowed range,
and calculated what actually zoom distance and tilt angle these
represented on the fly as needed.

This commit is an intermediate step toward implementing inertial
behavior for changing the Y-rotation, tilt angle, and zoom amount.
It will be easier to implement by changing the raw value over time,
rather than the percentage.