Flythrough intro camera moves to face direction of movement
This prevents scenarios where e.g. the camera is looking straight ahead,
but the camera is moving at a near horizontal angle to the right. This
leads to a strange "where am I going?" effect.

This removes some logic to make the transition to the bird's eye animation
smoother, which can be added back in a future commit.