Restoring transition in flythrough from intro -> bird's eye
This prevents the camera wildly rotating after reaching the city center
to move to the initial bird's eye position. Instead, it will rotate to
face the closest axis, and then travel on that axis for a short distance
before starting the main bird's eye animation. This is a much less jarring
experience.