More dynamic descents from/ascents into bird's eye view during flythroughs
Instead of a linear change from the original X-rotation to the target X-rotation,
the camera attempts to being the downward-looking camera position as long as
possible. This makes the animation look more interesting.

To do this, adding a new easing function. Calling it "steep easing" for lack of
a better name.