Flythrough intro rotates to correct angle when city not centered at 0,0
The intro animation always transitions from the initial flight to the city
center to an transitional bird's eye animation. The transitional animation
moves 5 axis-aligned map units away from the city center. This should always
result in a Y-rotation aligned to one of the 4 axes. However, the previous
logic always calculated the target to be 5 units away from the {0, 0}, and
not the city center. If the city wasn't centered at {0, 0}, this would result
in a diagonally rotated movement to the target.