Adjusting touch zoom/rotate-Y parameters
A slightly larger rotation angle is needed before Y-rotation will be
turned on, and the zoom delta must be a minimum amount before the zoom
amount will be changed. These changes serve to reduce the likelihood of
"accidental" input which can cause the touch controls to feel jittery
instead of smooth.