Using correct variables
The position/rotation coordinates should be set from the `initial` object,
and not the object-level coordinate variables. This happened to work when
animations were created "just in time", but causes problems if attempting
to create animations for later use (since the current vehicle position/rotation
in not necessarily what the future animation should use).