Capitalizing DOCTYPE declaration
After a quick search it seems like this tag should be treated case-insensitively,
so "doctype" and "DOCTYPE" should both be valid. However, since the lowercase
"doctype" seemed a little unusual and made me question it, it seems like capitalizing
it would look more normal and avoid a similar question-and-answer session in the future.