Preventing attempt to navigate to non-existent road intersection
The flythrough intro will always attempt to navigate to a road intersection
6 blocks away from the city center, in the closest cardinal direction from the
flight angle to the city center. However, there might not be a road intersection
in this position (for example, if the city center is close to a river bank).

Previously, this would result in an error once the flythrough reached the
phantom intersection, causing the flythrough to freeze and the page needing to
be refreshed.

Now, if this intersection doesn't exist, it will instead resort to a backup animation
at the city center. This animation is a little lame because it will result in extra
Y-rotation spinning to reach the target direction, but it's better than the page
crashing. Smarter behavior can probably be implemented in the future, but this at
least fixes the most immediate problem.