Mouse/touch panning moves world in world units, not arbitrary units
E.g., if you are looking straight down onto the scene and place your finger
on a building, the building will stay under your finger as you move your finger
around. This makes it feels more like you are directly manipulating the world.
Previously, at certain zoom distances/tilt angles the world would not as closely
track your finger movements, because the amount of movement was based on arbitrary
numbers that (through trial and error) happened to more or less match world units.

The amount of world movement (for a given zoom distance) should be the same
regardless of the mouse/finger position on the screen. I.e., moving 10 pixels up
near the top of the screen should correspond to the same world movement as moving
10 pixels up near the bottom of the screen. This might sound obvious, and is what
would be expected as a user. But a naive implementation might project rays from
the finger+camera position into the world to determine the world movement delta,
and if the camera is tilted at an angle, 10 pixels near the world horizon will
result in a larger world movement than 10 pixels near the bottom of the screen
(i.e. near the camera) due to perspective projection. This means that is some
cases very small pixel movements on the screen can result in giant jumps in the
world, which is obviously not desirable.

In order to make world movement deltas uniform regardless of the screen position,
the relationship between 1 pixel -> # of world units is calculated using a
simulated camera looking straight down on the center of camera orbit (regardless
of the actual camera position). This does means that moving 10 pixels will result
in the same world movement regardless of tilt angle - this may or may not be
desireable, since it can have the effect of making world movement seem sluggish
when the camera is tilted at a close to horizontal angle. This is something than
can be re-evaluated in the future, but despite that potential drawback this commit
seems like a good improvement over the previous behavior.