Touch zoom gesture zooms toward center of fingers, not center of screen
This allows you to zoom toward any arbitary location currently visible on
the screen, instead of only being able to zoom toward the current center.
It matches how Apple Maps touch zooming behaves, and feels like a better
UI overall.

When the camera is tilted close to a "looking straight forward" angle, this
gesture is hard to use in practice. Since the "center of fingers" world
position is determined by un-projecting the screen position -> world position,
touch positions close to/above the horizon can result in a center-of-zoom
that is very far away from the camera, which results in zooming toward a very
far away spot. This should be fixed in the future.