Extracting domain objects for mouse/touch points
These new domain objects better encapsulate the mapping of screen
pixels to normalized screen coordinates and world coordinates, as
well as relationships between multi-touches (distance between touches,
angle between touches, etc.) Not necessarily the be-all-end-all, but
this seems to incrementally make the touch navigation logic easier
to understand/modify.

Another benefit is that this helps make this code more easily tested
in the future.