Removing unnecessary duplicate test assertions
The removed tests are essentially the same assertions as other existing
assertions, and are only testing that methods work if passed either an
integer or a float. I think it's safe to assume that JavaScript will
handle either in the same way, so these extra tests just seem like noise.