Removing legacy fields for terrain bounds
And converting existing uses of the legacy fields to use the new fields.
The primary difference is that the new fields return the raw min/max based
on the scale, instead of clamping the value to an integer. This makes the
`Terrain` object more flexible, by allow the terrain to be scaled to any
arbitrary size, not just integer multiples.