Speeding up terrain mesh generation by re-using previous calculations
The triangles in the terrain mesh are generated from top to bottom,
in columns moving left to right. When calculating the various data for a
terrain triangle (is it bordered by roads? terrain height? water height?
etc.) the "bottom" values for the triangle above can be re-used as the
"top" values for the current triangle.

Somewhat surprisingly, re-using these values causes a ~30-50ms speedup in
the terrain mesh generation time in Safari, Chrome, and Firefox on my
computer. For example, in Safari it seems to bring terrain mesh generation
time down from ~200ms to ~150ms for terrain at the default scale.