Camera inertial velocity stops below a minimum threshold of movement
Otherwise, inertial velocity can continue forever, even when the movement
changes are so small that they won't cause any visible movement of the
camera on the screen. Since scene re-rendering is tied to camera movement,
this means that the scene will be indefinately re-rendered each animation
frame, even if it doesn't result in any actual differences. This is obviously
wasteful for performance + energy usage.

The minimum movement threshold value isn't calibrated to be exact. It was
chosen so that inertial velocity would stop a reasonable amount of time
after the camera movement is no longer affected, while also not stopping
inertial velocity too soon.