Initial implementation of neighborhood-based road network generator
Instead of generating a single "neighborhood" at the center of the
city, this generates several neighborhoods. This is very basic, and
currently results in fairly nonsensical results outside of the city
center, because the neighborhoods are not necessarily connected, and
buildings will only be built in the city center. But, it is a first
step.

The road network algorithm is essentially the previous road network
generation algorithm. This is to bootstrap things, but not necessarily
the actual algorithm that will be used long term.