Neighborhoods are connected by a road
The roads don't necessarily make sense, but this can be improved
in the future.

This introduces TerrainCandidateRoadNetwork, which has an interface
which allows it to be substituted for a RoadNetwork when using a
PathFinder instance. This allows finding the shortest path between
two points on the terrain, without having to have a road network
already in place.