Preventing excessive connector roads between neighborhoods
When building a road to connect a neighborhood with the rest of
the city, it will stop when reaching an existing road. In practice,
this causes neighborhoods that are near each other to share a single
road back to the city center, instead of each neighborhood building
it's own road (often close to neighboring connector roads). As a
side benefit, this causes road network generation to be noticeably
faster (typically multiple hundreds of milliseconds on my computer).

To help implement this, the "has reached the target?" check inside
`PathFinder` can now be parameterized.