Traditional road generation back-ported to neighborhood generator
This means that that previous "single giant road network in the
middle of the city" can be implemented using the new neighborhood-aware
generator by passing in the correct config values. (In this commit
the config is reverted back to those settings).

The legacy `RoadNetworkGenerator` object is removed in this commit,
so now the only way to generate a road network is to base it off of
the presence of one or more neighborhoods.