Camera no longer goes haywire when touch-zooming out from max zoom
Previously, this would cause the camera to jump to seemingly random
location:

1.) Using pinch-to-zoom, zoom all the way to max zoom (at any point
    in the scene).
2.) Release fingers
3.) Start another pinch-to-zoom, and zoom out.

Immediately upon zooming out, the camera would jump to the edge of
the terrain.

When zoomed all the way in, the denominator of a calculation would
be 0, resulting in Infinity. Since the camera position was based on
this value, it caused the camera to go as far to the edge of the world
as possible, instead of zooming out from the current position.