Adding alternate terrain eroder
This is much faster than the pre-existing terrain eroder, and seems
to produce more stable results. It needs more work before it is
practical to use for real cities though.

It adapts some ideas from this blog post:
http://ranmantaru.com/blog/2011/10/08/water-erosion-on-heightmap-terrain/

The implementation is simpler than what is described in that post,
which possibly could affect the quality of the results.