Adding shape generators for sheer ridge, sloped ridge on terrain
Similar to the "plateau" and "pyramid" shapes, these are intended
to be used with erosion for more realistic looking results.

The sloped ridge generator is not ideal, because the ends on the
ridge result in a sheer drop off. This is because a "distance of
point to line" formula is used to determine the terrain height,
instead of a "distance of point to line segment" formula.