City center chosen from centrality, terrain flatness
Previously, the city center location was determined by choosing a
position close to the center of the terrain that wasn't covered
by water. This could cause bad results for certain terrain (for
example, a city centered on the top of a mountain when there is a
lot of flat terrain nearby).

Now, the flatness of the terrain will be taken into account. This
means that a location can be chosen if it is not especially close
to the terrain center, but if it is flatter than other more central
locations. Flatness is weighted more heavily than centrality.

The flatness calculation only takes the immediate city center into
account, which still means a city might be sited on flat terrain
adjacent to a mountain, etc. However, from testing on some example
terrain this still seems like a big improvement.