Neighborhoods build themselves near city center
Previously, neighborhoods (other than the city center neighborhood)
would place themselves in a random non-water location. Now, they
will try to place themselves in suitable flat terrain close to the
city center, but not too close to other neighborhoods.

The end result is a city that will expand out from the center, but
follow the terrain. For example, if a city is centered in an
irregularly shaped valley, the city will expand to fill the valley
before trying to climb up the sides of the surrounding mountains.
(More or less. The config for weighting flatness vs. location can
probably be improved).