Neighborhoods look for flatter terrain
Previously, they only looked for flatness in a 3x3 square centered
on the center point. Now they look for flatness in a 9x9 square.
This seems to give better results by making it less likely a
neighborhood will be built on steep terrain that happens to have a
small flat part.

As a result, the boost for flatness in the overall score is also
being reduced, to reduce the likelihood of placing neighborhoods in
very flat terrain far from the city center. This has the effect of
causing the city to be built more closely around the center.