Better range definition for random terrain corner heights
Previously, the random corner heights to initialize the midpoint
displacement or diamond-square algorithms were chosen from a range
between 0-n. Now, they can be chosen from an configured range.

This allows the range to be evenly centered around 0, which makes
it easier to treat 0 as "sea level" in the future.