`RoadNetwork` internally stores distance between each edge
Currently, the distance is hard-coded to 1.0, but in the future it
could be set to an actual value.

The reason for doing this is to allow the "terrain that acts like
a road network for path-finding purposes" (i.e. the
`TerrainCandidateRoadNetwork` object) to treat cross-river connections
as having a non-1.0 length. This will help fix the broken logic for
building bridges to connect neighborhoods.