Experimentally adding ability to add random pyramids to terrain
By default, the number of pyramids is set to 0, so it doesn't have
any effect.

This is intended to be used with erosion as a way to simulate more
interesting, mountainous terrain that has flat valleys suitable for
building a city. Pyramid heights are limited in the center of the
terrain so that mountains are less likely to occur on top of rivers.