Neighborhoods can build bridges to connect themselves to rest of city
Previously, neighborhoods could build a "road over water", which was
treated like a normal road. This led to weird artifacts such as bridges
the curved across the water, buildings built in water, etc.

As part of doing this, the probability that a bridge will be built in
a valid location is now configurable.