Road simplification takes all directions into account
Previously it only attempted to remove west->east or north->south
roads. Now it also attempts to remove east->west and south->north
roads.

However, this algorithm is flawed. The reason is that removing a
road in one direction can open up an opportunity to remove a road in
a different direction. However, if that direction has already been
attempted, there will not be any way to exploit that and the newly
removeable road will remain.

For example, if a north->south road is removed, that could open up an
opportunity for an adjacent west->east road to be removed. However,
since west->east road removal always happens before north->south road
removal, the west->east road will never be removed.

You could work around this by running road removal multiple times,
but that seems gross, and is also not guaranteed to fully remove all
redundant roads if you run it for a fixed number of iterations.

Despite the problems with this algorithm, adding a full implementation
to help develop a more robust alternative.