Road network simplification runs until no roads left to remove
One pass of the current algorithm is not guaranteed to remove
all of the possible extra roads that can be removed. However, a
pass can open up opportunities to remove roads in a subsequent
pass. The algorithm is now run until a pass occurs in which no
roads were removed, indicating that all possible extra roads have
been removed. (As long as the algorithm works correctly).

The tradeoff is that the road simplification process can take
longer than before.