Neighborhoods that can't be connected to city don't cause error
If the terrain is extra hilly, it's possible that some neighborhoods
will be chosen in locations where they can't be connected to the rest
of the city. Rather than blow up in these cases, these neighborhoods
are silently dropped.

It would be better to avoid generating neighborhoods in places where
they can't be connected, but in practice it might not always be possible
to satisfy the "build in a suitable location" and "build in a place that
can be connected to the rest of the city" constraints for the full set
of desired neighborhoods.