More accurate touch points; debug mode that shows touch points
Previously, the touch point->world position calculations did not
take terrain height into account. If you touch on a hill, the
expected behavior is that the hillside will be treated as the
hillside, and not some point in the distance that would be the
correct location if the terrain were completely flat.

This implementation uses the Three.JS raycaster to perform the
touch point->world position calculation, which is slow and causes
a noticeable degredation in touch performance on mobile. Maybe a
better algorithm would be faster in the future.

This also adds a debug mode where the touch points are indicated
on the screen with special meshes. Ideally the debug mode should
be added in a separate commit, but these two feature are intertwined
and I am too lazy to separate them into separate commits.