Preventing out-of-memory error when creating new city
Previously, the process of creating a new city meant generating
the new city, and then replacing the previous city with it.
However, this meant that there were temporarily two cities in
memory at the same time. This seemed to cause out-of-memory errors
on devices with limited memory, such as a phone.

Now, the original city is destroyed before creating the new city,
which seems to prevent this problem (at least, as long as the
device has enough memory to store a single city in the first place).