Better implementation of adding pyramid to terrain
The shape added is now an actual pyramid, rather than a cone. To
simplify things, the shape of the pyramid base must now be a square,
instead of a rectangle, and the base length must be an even number.
In the future, it might be good to make it more flexible, but this
provides a better base to work off of.

In addition, adding initial tests for this functionality.