Faster, more accurate calculation of screen touch->world position
Like the previous implementation that used the Three.JS Raycaster,
this implementation will correctly place the touch point on the
correct terrain position based on the terrain height, instead of
assuming a flat Y plane. However, this implementation is much faster
(<=1ms vs ~25-30ms on my computer). It also seems to be a little
more accurate - the Raycaster method seemed to not always hit the
exact target for some reason.

This implementation casts a ray from the current camera position in
the direction of the y-plane, and checks at a regular internal to
see if the ray is below the terrain height at the current location.
If so, that means the ray has intersected the terrain, and the
current ray location should be the world touch position. It is much
faster than the Raycaster implementation because it only needs to
check for intersection at X/Z coordinates along the ray path, instead
of every triangle in the scene.

This current has a known limitation that it will fail if the touch
point is outside of the world bounds - that should be fixed in a
future commit.