"Zoom toward target" takes terrain elevation into account
Previously, the "zoom toward target" logic assumed the Y coordinate
was 0. This meant that when zooming toward any terrain with a height
greater than 0, your zoom would not necessarily go where you would
expect. The higher the terrain, the greater the discrepency.

A known issue with this commit is that zooming won't stop once you
intersect the terrain - it will keep going until the zoom is at the
maximum level, which can cause the zoom to "overshoot" the target
and keep going. You won't go underground, because that is not
allowed, but your final X/Z coordinate will be beyond the target
of zoom location.

This also adds a debug marker for the X/Z location of the zoom target
end point.