Touch points outside of terrain bounds don't cause an error
This makes it easier to pan the world without having to touch the
terrain at all times.

There is still a failure case where the ray will never intersect with
the terrain or the XY plane: if the tilt angle is close to 0 (i.e. level
with the ground) and the touch point is in the sky, then the ray will
point away from the ground and never intersect with the XY plane. This
will cause an infinite loop, and should be fixed in a future commit.