Same edge object shared by both nodes of a road segment
Prevents the edge between the two nodes from having different data
depending on which node is queried (not because there is a known
existing error case, but for defensiveness), and to also possibly
reduce memory usage by halving the number of edge objects that
need to be stored. (I haven't profiled memory usage, but this still
seems like a safety improvement even if it doesn't have much practical
impact on memory usage).