Improving performance of zoned block generation
This avoids evaluating terrain coordinates outside the bounds of
the road network, since those coordinates are guaranteed to not
be suitable (since all lots must be placed next to a road).

When generating a city with max number of neighborhoods, on flat
terrain with no river, this seems to increase the speed of zoned
block generation by about 10-20ms on my computer in Safari and Chrome.