More detailed performance logging of city generation time
This adds logging for the amount of time needed to resize the
three.js renderer, which happens each time the scene geometry is
replaced. This takes an unexpected amount of time (multiple
hundreds of milliseconds?!).

This also logs the total time needed to complete the entire
process, to help make it easier to detect hidden performance
problems in the future. (I.e. if there's a big discrepency
between the total time and the sum of the individual component
times, that implies there's an un-measured part of the process
that is taking a long time).