OrbitalCamera no longer has it's own reference to Terrain object
When `OrbitalCamera.syncToCamera()` is called, the `Terrain` object
must be supplied as part of the function call. This reduces the
amount of state in `OrbitalCamera`, and reduces the number of places
state must be updated when a new city is generated. (`syncToCamera()`
is the only place that `OrbitalCamera` used the `Terrain` reference).