SceneView no longer has a reference to a Terrain instance
The previous reason was so that `NavigationTouchController` had
access to the current `Terrain` instance; it already had access
to the current `SceneView` so it was convenient. However, this
didn't really make sense with respect to `SceneView`s intended
purpose. Although this now requires `NavigationTouchController`
state to be updated when a new city is generated, the benefit of
`SceneView` having a more clear responsibility seems worth it.