Adding basic zoom in/out buttons
Intended for non-touch environments. The future benefit is that
dedicated zoom in/out buttons will allow automatically stopping the
zoom when the camera hits terrain, which doesn't really work very
well with a slider. This will allow the non-touch zoom behavior to
match the behavior of touch controls. However, in this implementation,
the buttons aren't terrain-aware, and simply zoom in/out in the
same way as the slider.