Initial implementation of use two finger drag to pan the world
This doesn't play nicely with the pinch zoom gesture yet. To make
it a little less broken, the amount of finger movement needed to
start a pinch zoom gesture is increased. Otherwise, trying to pan
is very glitchy due to conflict with the zoom gesture. This is just
a temporary workaround, the real fix is to allow the two gestures
to properly co-exist.