Pinch-zoom implementation based on zoom distance (not percentage)
This is largely a re-implementation of the pinch-zoom gesture. It
seems more conceptually simple. Instead of basing the zoom on the
"percentage of zoom toward the maximum", it is based on the actual
zoom distance. This seems more direct, and makes the logic less
reliant on knowledge of details like the orbital camera's min/max
zoom distances. It also removes the need to know the specific
center X/Z coordinates of the orbital camera at the point where the
zoom started.

This no longer implements the "decrease zoom amount the more you
zoom in" behavior, in favor of a constant zoom amount. The previous
behavior should be re-implemented in a future commit, but is left
out of this commit to simplify the initial implementation.