Orbital camera can only have zoom distance set, not zoom percentage
This requires updating `NavigationController` to set all zoom values
in terms of zoom distance, not "percentage of max zoom level". This
is in line with removing the manual zoom slider in the future, in
favor of only allowing using the "zoom in"/"zoom out" buttons.

This allows simplifying `OrbitalCamera` and removing some code.