Extracting `MapCamera`, to encapsulate rotate/zoom/pan gestures
`MapCamera` wraps an `OrbitalCamera`, and allows implementing
actions such as "zoom toward point" or "rotate around point".

Previously, this logic was a part of `GestureProcessor`. However,
this coupled detected of touch gestures with the internal logic
of abstract actions like "zoom toward point". It also made it
more difficult for mouse/touch interfaces to use the same camera
logic behind the scenes.