`MapCamera` always has a center-of-action
Now it should never be `undefined`. This should make it easier to
implement desktop/mouse controls in terms of `MapCamera` in the future.

This requires changing some of the logic for determining the current
touch gesture, in particular to determine the gesture not solely
based on the current touch inputs, but also the previous in-progress
gesture.