Better interface for setting pan distance in `MapCamera`
It doesn't seem like `MapCamera` should be responsible for translating
raw screen movements into a pan distance amount. Not sure that
`GestureProcessor` is the ideal place either, but this seems like a
step in the right direction.

A benefit of this is that the inertial panning implementation in
`MapCamera` is more straightforward. This also makes `MapCamera.pan()`
more consistent with that other similar methods (e.g.
`zoomTowardCenterOfAction`) in that it now takes a raw amount of
movement as the argument.