Inertial zooming/azimuth rotation happens around correct point
Previously, the `MapCamera` center-of-action was reset to the
underlaying `OrbitalCamera` center point whenever there were no
touch points present. However, this meant that inertial zooming
and azimuth rotation would not occur around the intended center of
action.

Since the center of action is reset when a new gesture is started,
it doesn't seem like this should have any drawbacks.