More accurate center of action for two finger touch gestures
The center of action is now determined from the _screen_ midpoint
projected into the world, rather than the midpoint of the the two
touch points in world coordinates.

This results in a more expected center of action, since presumably
the user expectation is that it will be "the point in the world
that is visually between my two fingers".

The problem with using world coordinates to calculate the midpoint
is that when the tilt angle is close to level with the ground,
the visual midpoint is not necessarily the same as (or even close
to) the world midpoint. For example, if one of the touch points
is close to the horizon, and another closer to the foreground, the
world midpoint will visually be closer to the horizon touch point
than the foreground touch point, instead of visually halfway
between them. (Due to perspective). This results in zooming toward
or rotating around an unintended location.