Amount of zoom in/out button zoom proportional to current zoom amount
This causes the amount of zoom to decrease as you get closer to the
XZ plane. This allows zoom close to the surface to be sensitive, while
not being overly sluggish when zoomed a long way out.

This only affects the manual zoom in/out buttons, and not the pinch
zoom touch gesture (although that should be changed to be proportional
as well in the future).