Rescaling 3D scene to have same scale as raw terrain/road/etc. data
In practice, this means reducing the scale of the 3D scene by 12, since
that was the amount that the 3D scene was scaled up from the underlaying
raw terrain/road/lot/building data. As long as the scale of all meshes
and camera movements are scaled down by the same amount, this should be
visually indistinguishable from the old scale.

This removes the need to convert between "map" and "scene" coordinates,
since they are now 1:1. Although not changed in this commit, this means
that in the future usage of `CityTour.Coordinates` can be removed, because
the relevant methods (e.g. `mapXToSceneX`) now return the same value they
are called with.

Besides the benefit of simplification, this allows sizes for the Y dimension
to be in the same scale as the X/Z scales. For example, previously the X/Z
dimensions of a building were defined in "map" coordinates, but the height
was expressed in "scene" coordinates. It is simpler and less confusing to
use the same scale for the X/Y/Z axes at all times.

The different scales date back to the earliest days of the project, when
A.) I had less of an idea what I was doing and B.) the project was much
simpler. As far as I can tell, in the immediate term there aren't any
drawbacks other than the effort/bug risk of making changes across the whole
codebase. If differing scales are desired in the future for some reason,
the capability can be added back in, but without the need to maintain the
functionality until it is actually needed.

There are undoubtedly some changes that have been missed in this commit, but
from initial testing this basically seems to work.