Driving portion of flythrough moves closer to the ground
This sets the driving distance off the ground to be 1/12th of the
original distance off the ground, in line with the "great re-scaling
of 2018". (See commit f43e94b7f4c5b0540b3addb282ad219a47db3b1b).

Now that the camera "near" value is updated to be compatible with
the new 1/12th scale (see 59653bda0fa852c103a01a686794504f302a637e),
the distance off the ground can be set to the proper value without
causing the underside of the world to be seen excessively when moving
on sloped roads.