Pinch zoom speed slows the more zoomed in you are
This matches the behavior of the manual zoom in/out buttons, and
makes it easier to zoom into the world at a close zoom level without
the camera being too jumpy.

This should ideally also be proportional to the change in the distance
between the two touch points, but this is a step in the right direction.