More accurate screen touch -> world position calculation
This makes the calculation more accurate by moving along the ray in
smaller steps toward the target. There's a tradeoff between a more
accurate calculation (i.e. smaller step size) and performance. This
new step size seems to provide a noticeable improvement in accuracy
(tested manually by observing the location of the "center of action"
debug marker on the screen) while not causing a meaningful degredation
in performance.