More efficient way to avoid placing neighborhoods too closely
Each terrain point now has the concept of a "close neighbor penalty".
This penalty starts at 0 for every terrain point, meaning that there
is no penalty. When a neighborhood is built, terrain points near the
neighborhood center are set to have an infinite penalty, preventing
a neighborhood from being built there.

The benefit of this approach is that the distance to the closest
neighborhood no longer needs to be calculated for every terrain point
on each neighborhood placement iteration - only the points near the
new neighborhood. Since the penalty is essentially binary, once the
value is set to infinity is will always be enabled regardless of any
other neighborhoods that get built.