More easily tuned tradeoff between flatness, centrality
The terrain flatness around a candidate neighborhood center and the
closeness to the center of the terrain are now given explicit weights
in the calculation to determine the best neighborhood site. The two
values are now also on the same scale (0 to 1). This should make it
easier to experiment with different weights to see how it affects
neighborhood generation.

To start off, flatness is given a higher weight than centrality,
since this seems to result in more realistic cities that avoid hills
unless very close to the city center.