Manual rotation revolves around terrain point at center of screen
Instead of the intersection with the XZ plane from a vector extending
from the screen center.

The intended benefit is to make the center of rotation more expected
when the terrain is hilly. Conceptually, this simulates a touch
rotation around the exact center of the screen (because touch rotations
already take the terrain height into account when calculating screen
touch coordinate -> world coordinate).

This is a little experimental, and will live with it a bit to see how
well it works. But so far, it seems like an improvement?