Allowing rays moving away from XZ plane to intersect with terrain
Previously, if the ray had a positive Y direction (i.e., the ray
is moving away from the XZ plane because we assume we are always
above it), screen coordinate -> world terrain coordinate projection
would short-circuit and act as if there was no intersection with
the terrain.

The benefit was that this prevented infinite loops when e.g. the
screen coordinate was in the sky. However, it prevented legitimate
hit detection is certain cases, such as when right in front of
a mountain with the camera looking straight toward the horizon.
A screen coordinate at the top of the screen could represent
terrain in this case, not sky.

This approach of approximating "inifinite loop" with "iterated too
many times" doesn't seem ideal, but this seems like a reasonable
work around until a better approach is discovered. Performance-wise
it seems fine?